GRUB + Example File + Project Requirements + Programs
I created a script to modify the bootloader.

I created a script to create and set the proper permissions of the example file.

I then created a script to install all of the other projects required commands.

I then created a script to display all of the output that the program is asking for.